♻️ refactor: vote.ts 중복 파일 삭제 및 수정
