+++ src/js/chat/vote.ts
@@ -9,8 +9,14 @@
  * - 동률일 경우 처형하지 않음
  */
 
-import { getPlayerList, setPlayerList } from '../lib/store';
+import {
+    getPlayerList,
+    initVote,
+    setLiveOrDiePlayer,
+    setPlayerList,
+} from '../lib/store';
 import { sendMsg, socket, type ChatMessage } from '../lib/yongchat';
+import { showText } from './chatting';
 
 /**
  * 플레이어가 다른 플레이어에게 투표하는 함수
@@ -52,65 +58,95 @@ function addVote(targetId: string) {
     console.log('득표', targetId);
 }
 
-/**
- * 투표를 처리하고 결과를 계산하는 함수
- * @param {string} targetId - 투표 대상 플레이어의 ID
- */
-// function addVote(targetId: string) {
-//     // 현재 게임의 전체 플레이어 목록 조회
-//     const playerList = getPlayerList();
+// 나의 유저 ID (현재 로그인된 유저가 user1라고 가정함) - 자기 가진을 투표할 수 없도록
+const USER_ID = 'user1';
 
-//     // 투표 대상 플레이어의 득표수 증가
-//     playerList[targetId].vote++;
+const HIDE_CLASSES = ['invisible', 'opacity-0']; //숨기기
+const SHOW_CLASSES = ['visible', 'opacity-100']; //보이기
 
-//     // 전체 투표수 계산을 위한 변수 초기화
-//     let totalVote = 0;
-//     const playerCount = Object.keys(playerList).length;
+// 현재 선택된 대상 유저ID 저장
+let selectedUserId: string | null = null;
 
-//     // 모든 플레이어의 득표수를 합산
-//     for (const playerId in playerList) {
-//         const player = playerList[playerId];
-//         totalVote += player.vote || 0;
-//     }
+// 모든 프로필 요소 선택
+const profileElements = document.querySelectorAll<HTMLElement>('.profile');
 
-//     // 변경된 플레이어 정보를 저장소에 업데이트
-//     setPlayerList(playerList);
+profileElements.forEach(profile => {
+    //모든 유저 프로필에 설정된 data-user 가져오기
+    const targetUserId = profile.dataset.user;
 
-//     // 투표 진행 상황 로깅
-//     console.log('득표', targetId);
-//     console.log('현재까지 투표수', totalVote);
-//     console.log('목표 투표수', playerCount);
+    // 유저 id가 자기 자신이면 클릭 이벤트 적용 안됨
+    if (!targetUserId || targetUserId === USER_ID) return;
 
-//     // 모든 플레이어가 투표를 완료했는지 확인
-//     if (totalVote === playerCount) {
-//         console.log('투표 종료', playerList);
+    profile.addEventListener('click', () => {
+        // 프로필 안에 있는 .vote-icon 선택
+        const voteIcon = profile.querySelector<HTMLElement>('.vote-icon');
+        if (!voteIcon) return;
 
-//         // 최다 득표수 계산
-//         const maxVote = Math.max(
-//             ...Object.values(playerList).map(player => player.vote),
-//         );
+        // 이미 선택된 유저를 다시 클릭하면 → 투표 취소
+        if (selectedUserId === targetUserId) {
+            voteIcon.classList.remove(...SHOW_CLASSES); //아이콘 숨기기
+            voteIcon.classList.add(...HIDE_CLASSES); //아이콘 숨김
+            selectedUserId = null; //선택 상태 초기화
+            // console.log('투표 취소');
+            return;
+        }
 
-//         // 과반수 이상 득표 여부 확인
-//         if (maxVote >= playerCount / 2) {
-//             // 최다 득표자 목록 추출
-//             const maxVotePlayers = Object.values(playerList).filter(
-//                 player => player.vote === maxVote,
-//             );
+        // 다른 유저가 이미 선택되어 있으면 그 유저 아이콘 숨기기(중복 안되도록)
+        if (selectedUserId) {
+            // 이전에 선택한 유저 프로필 아이콘 찾기
+            const prevIcon = document.querySelector<HTMLElement>(
+                `.profile[data-user="${selectedUserId}"] .vote-icon`,
+            );
+            prevIcon?.classList.remove(...SHOW_CLASSES); //보이던 아이콘 숨김 처리
+            prevIcon?.classList.add(...HIDE_CLASSES);
+        }
 
-//             // 최다 득표자가 한 명인 경우 (동률이 아닌 경우)
-//             if (maxVotePlayers.length === 1) {
-//                 // 해당 플레이어를 처형 대상으로 지정
-//                 setLiveOrDiePlayer(maxVotePlayers[0]);
-//                 // 처형 대상자 발표
-//                 showText({
-//                     action: 'chat',
-//                     nickname: '사회자',
-//                     msg: `${maxVotePlayers[0].nickName}님이 총 ${maxVote}표를 받아서 사형대에 올랐습니다. 찬반 투표를 진행해 주세요`,
-//                 });
-//             }
-//         }
+        // 현재 클릭한 유저의 아이콘 표시
+        voteIcon.classList.remove(...HIDE_CLASSES); //숨기기 제거 - 현재 클릭한 유저의 아이콘을 보이게함
+        voteIcon.classList.add(...SHOW_CLASSES); //보이기 클래스 추가
+        selectedUserId = targetUserId; //현재 선택된 유저 id 갱신
+        // console.log('투표 대상:', selectedUserId);
+    });
+});
 
-//         // 투표 상태 초기화
-//         initVote();
-//     }
-// }
+// 최다 득표자 처리 함수
+export function handleVoteResult() {
+    const playerList = getPlayerList(); // 플레이어 리스트를 가져옴
+    const playerCount = Object.keys(playerList).length; // 살아있는 플레이어의 수를 체크
+
+    // 최다 득표수 계산
+    const maxVote = Math.max(
+        ...Object.values(playerList).map(player => player.vote || 0),
+    );
+
+    if (maxVote >= playerCount / 2) {
+        const maxVotePlayers = Object.values(playerList).filter(
+            player => (player.vote || 0) === maxVote,
+        );
+
+        if (maxVotePlayers.length === 1) {
+            // 단일 최다 득표자만 있을 경우
+            setLiveOrDiePlayer(maxVotePlayers[0]);
+            showText({
+                action: 'chat',
+                nickname: '사회자',
+                msg: `${maxVotePlayers[0].nickName}님이 총 ${maxVote}표를 받아서 사형대에 올랐습니다. 찬반 투표를 진행해 주세요`,
+            });
+        } else {
+            // 동률일 경우 처형 없음
+            showText({
+                action: 'chat',
+                nickname: '사회자',
+                msg: `투표 결과 동률이 발생하여 처형 대상자가 없습니다.`,
+            });
+        }
+    } else {
+        showText({
+            action: 'chat',
+            nickname: '사회자',
+            msg: `과반수 득표자가 없어 처형 대상자가 없습니다.`,
+        });
+    }
+
+    initVote(); // 다음 투표를 위해 초기화
+}
